+++ Package.swift
@@ -3,6 +3,8 @@
 /* Package.swift
  Version: 1.1.0
 
+ For iOS and macOS only. Use Stars to adopt for the specifics you need.
+
  Created by Mikhail Zhigulin in 7533.
 
  Copyright © 7531 - 7533 Mikhail A. Zhigulin of Novosibirsk
@@ -15,7 +17,7 @@
  All rights reserved.
 
  Abstract:
- Package manifest for an App component.
+ Package manifest for ConsolePerseusLogger.
 */
 
 import PackageDescription